+++ relayer/cmd/run/run.go
@@ -4,6 +4,7 @@ import (
 	"github.com/snowfork/snowbridge/relayer/cmd/run/beacon"
 	"github.com/snowfork/snowbridge/relayer/cmd/run/beefy"
 	"github.com/snowfork/snowbridge/relayer/cmd/run/ethereum"
+	"github.com/snowfork/snowbridge/relayer/cmd/run/execution"
 	"github.com/snowfork/snowbridge/relayer/cmd/run/parachain"
 	"github.com/spf13/cobra"
 )
@@ -19,6 +20,7 @@ func Command() *cobra.Command {
 	cmd.AddCommand(parachain.Command())
 	cmd.AddCommand(ethereum.Command())
 	cmd.AddCommand(beacon.Command())
+	cmd.AddCommand(execution.Command())
 
 	return cmd
 }
