+++ relayer/relays/beacon/config/config.go
@@ -25,7 +25,6 @@ type SourceConfig struct {
 
 type BeaconConfig struct {
 	Endpoint                string `mapstructure:"endpoint"`
-	FinalizedUpdateEndpoint string `mapstructure:"finalizedUpdateEndpoint"`
 	Spec                    Spec   `mapstructure:"spec"`
 	ActiveSpec              string `mapstructure:"activeSpec"`
 }
